Move validation statistics out
