+++ great_expectations/validator/validator.py
@@ -17,7 +17,6 @@ from typing import (
     Callable,
     Dict,
     List,
-    NamedTuple,
     Optional,
     Sequence,
     Set,
@@ -81,6 +80,9 @@ from great_expectations.validator.validation_graph import (
     ExpectationValidationGraph,
     MetricEdge,
     ValidationGraph,
+)
+from great_expectations.validator.validation_statistics import (
+    calc_validation_statistics,
 )
 
 logger = logging.getLogger(__name__)
@@ -153,14 +155,6 @@ class ValidationDependencies:
         Returns "MetricConfiguration" dependency objects specified.
         """
         return list(self.metric_configurations.values())
-
-
-class ValidationStatistics(NamedTuple):
-    evaluated_expectations: int
-    successful_expectations: int
-    unsuccessful_expectations: int
-    success_percent: float | None
-    success: bool
 
 
 @public_api
@@ -2058,29 +2052,3 @@ class BridgeValidator:
                 **self.init_kwargs,
                 **self.batch.batch_kwargs.get("dataset_options", {}),
             )
-
-
-def calc_validation_statistics(
-    validation_results: list[ExpectationValidationResult],
-) -> ValidationStatistics:
-    """
-    Calculate summary statistics for the validation results and
-    return ``ExpectationStatistics``.
-    """
-    # calc stats
-    evaluated_expectations = len(validation_results)
-    successful_expectations = len([exp for exp in validation_results if exp.success])
-    unsuccessful_expectations = evaluated_expectations - successful_expectations
-    success = successful_expectations == evaluated_expectations
-    try:
-        success_percent = successful_expectations / evaluated_expectations * 100
-    except ZeroDivisionError:
-        success_percent = None
-
-    return ValidationStatistics(
-        successful_expectations=successful_expectations,
-        evaluated_expectations=evaluated_expectations,
-        unsuccessful_expectations=unsuccessful_expectations,
-        success=success,
-        success_percent=success_percent,
-    )
